requirements.txt
diffusers
einops
gin-config
matplotlib
moviepy
numpy
onnx
pandas
polars
pyarrow
pytorch-lightning
tensorboard
tensorboardX
tensorrt
timm
torcheval
transformers
wandb
